+++ susidko.fedor/P3/funcs.hpp
@@ -1,9 +1,10 @@
+#include <cstddef>
 #ifndef FUNCS_HPP
 #define FUNCS_HPP
 
 namespace susidko
 {
   char * replaceUpLow(char * src, int add);
-  int countDiffLetters(char * src);
+  size_t countDiffLetters(char * src);
 }
 #endif
